unified buttons (#176)
