+++ components/Button/Button.jsx
@@ -1,5 +1,8 @@
 import PropTypes from 'prop-types';
 import cx from 'classnames';
+import Router from 'next/router';
+
+const handleLink = (url) => window.open(url, '_blank');
 
 const Button = ({
   onClick,
@@ -7,24 +10,36 @@ const Button = ({
   type,
   isSecondary,
   className,
+  route,
+  internalQuery,
   ...otherProps
-}) => (
-  <button
-    className={cx(
-      'govuk-button',
-      {
-        'govuk-button--secondary': isSecondary,
-      },
-      className
-    )}
-    data-module="govuk-button"
-    onClick={onClick}
-    type={type}
-    {...otherProps}
-  >
-    {label}
-  </button>
-);
+}) => {
+  const isExternal = route && route.includes('https://');
+  const handleClick = () => {
+    onClick?.();
+    route &&
+      (isExternal
+        ? handleLink(route)
+        : Router.push(internalQuery ? `${route}${internalQuery}` : `${route}`));
+  };
+  return (
+    <button
+      className={cx(
+        'govuk-button',
+        {
+          'govuk-button--secondary': isSecondary,
+        },
+        className
+      )}
+      data-module="govuk-button"
+      onClick={handleClick}
+      type={type}
+      {...otherProps}
+    >
+      {label}
+    </button>
+  );
+};
 
 Button.propTypes = {
   onClick: PropTypes.func,
@@ -32,6 +47,8 @@ Button.propTypes = {
   type: PropTypes.string,
   isSecondary: PropTypes.bool,
   className: PropTypes.string,
+  route: PropTypes.string,
+  internalQuery: PropTypes.string,
 };
 
 export default Button;
